feat: add CardPieChart component
- Creates the CardPieChart component

- Handles the logic for building the Expenses by Category chart
